feat(hackathon): details page added fix(Modal z-index issue fixed):
