+++ src/App.tsx
@@ -44,16 +44,16 @@ import TaskBulkImport from "./modules/Dashboard/modules/Tasks/TaskBulkImport";
 import ZonalDashboard from "./modules/Dashboard/modules/zonalDashboard/zonaldashboard";
 import DistrictDashboard from "./modules/Dashboard/modules/districtDashboard/districtdashboard";
 import Hackathon from "./modules/Dashboard/modules/Hackathon/User/Hackathon";
-import HackathonCreate from "./modules/Dashboard/modules/Hackathon/Admin/pages/HackathonCreate";
+import HackathonCreate from "./modules/Dashboard/modules/Hackathon/pages/HackathonCreate";
 import KKEMLanding from "./modules/KKEM/modules/KKEMLanding";
 import KKEMAuth from "./modules/KKEM/modules/KKEMAuth";
 import RankCard from "./modules/Embeds/modules/RankCard/pages/RankCard";
-import HackathonEdit from "./modules/Dashboard/modules/Hackathon/Admin/pages/HackathonEdit";
 import LearningCircle from "./modules/Dashboard/modules/LearningCircle/pages/LearningCircle";
 import CreateCircle from "./modules/Dashboard/modules/LearningCircle/pages/CreateCircle";
-import { HackathonOrganizers } from "./modules/Dashboard/modules/Hackathon/Admin/pages/HackathonOrganizers";
+import { HackathonOrganizers } from "./modules/Dashboard/modules/Hackathon/pages/HackathonOrganizers";
 import FindCircle from "./modules/Dashboard/modules/LearningCircle/pages/FindCircle";
 import { LearningCircleLandingPage } from "./modules/Dashboard/modules/LearningCircle/pages/LearningCircleLandingPage";
+import HackathonDetails from "./modules/Dashboard/modules/Hackathon/pages/HackathonDetails";
 
 const router = createBrowserRouter([
     // Add redirect from '/' to '/login'
@@ -195,17 +195,29 @@ const router = createBrowserRouter([
                         element: <HackathonCreate />
                     },
                     {
-                        path: "hackathon/edit/:id",
-                        element: <HackathonEdit />
+                        path: "hackathon/details/:id",
+                        element: <HackathonDetails />
                     },
                     {
-                        path: "hackathon/test/:id",
+                        path: "hackathon/organizers/:id",
                         element: <HackathonOrganizers />
                     },
-                    { path: "learning-circle", element: <LearningCircleLandingPage /> },
-                    { path: "learning-circle/details", element: <LearningCircle /> },
-                    { path: "create-circle", element: <CreateCircle /> },
-                    { path: "find-circle", element: <FindCircle /> },
+                    {
+                        path: "learning-circle",
+                        element: <LearningCircleLandingPage />
+                    },
+                    {
+                        path: "learning-circle/details",
+                        element: <LearningCircle />
+                    },
+                    {
+                        path: "learning-circle/find-circle",
+                        element: <FindCircle />
+                    },
+                    {
+                        path: "learning-circle/create-circle",
+                        element: <CreateCircle />
+                    }
                 ]
             }
         ]
