+++ src/modules/Dashboard/modules/Hackathon/pages/styles.css
@@ -1,0 +1,185 @@
+
+.profileDash {
+	position: relative;
+	top: 2rem;
+	margin-left: 3rem;
+	display: flex;
+	gap: 2rem;
+	width: calc(100% - 6rem);
+}
+
+.content {
+	width: 100%;
+	display: flex;
+	flex-direction: column;
+	gap: 4rem;
+}
+
+.banner {
+	width: 150rem;
+	border-radius: 1rem;
+}
+
+.banner img {
+	width: 100%;
+	height: 100%;
+	object-fit: contain;
+	border-radius: 1rem;
+}
+
+.event__card {
+	width: 100%;
+	min-width: 20rem;
+	display: flex;
+	flex-direction: column;
+	justify-content: space-between;
+	gap: 1rem;
+}
+
+.apply__Card {
+	display: flex;
+	flex-direction: column;
+	gap: 1rem;
+	padding: 1.5rem;
+	border-radius: 1rem;
+	box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
+}
+
+.apply__Card .date {
+	display: flex;
+	align-items: center;
+}
+
+.apply__Card span {
+	display: block;
+	width: 2px;
+	height: 100%;
+	margin-right: .5rem;
+	background-color: #456FF6;
+}
+
+.apply__Card .close {
+	background-color: #F0F4FF;
+	padding: .5rem;
+	border-radius: .3rem;
+}
+
+.apply__Now {
+	border-radius: .3rem;
+	padding: .8rem;
+	background-color: #456FF6;
+	color: white;
+	border: none;
+}
+
+.event__card .social__links {
+	display: flex;
+	justify-content: center;
+	align-items: center;
+	gap: 2.5rem;
+	padding: 1rem;
+	box-shadow: 0px 4px 10px rgba(0, 0, 0, 0.1);
+	border-radius: 1rem;
+	font-size: 1.5rem;
+}
+
+.desc {
+	width: 46rem;
+	margin-left: 3rem;
+	background: white;
+	border-radius: 10px;
+	box-shadow: 0px 0px 10px 1px rgba(0, 0, 0, 0.1);
+	padding: 1.5rem 2rem;
+	display: flex;
+	justify-content: center;
+	flex-direction: column;
+	gap: 1rem;
+	margin-bottom: 2rem;
+}
+
+.description {
+	display: flex;
+	align-items: center;
+	gap: 1rem;
+}
+
+.hackprofile {
+	display: block;
+	border-radius: 100%;
+	height: 4rem;
+	width: 4rem;
+	object-fit: cover;
+	background-color: #000;
+}
+
+.love {
+	width: 2.5rem;
+	height: 2.5rem;
+	border-radius: 50%;
+	display: block;
+	background-color: pink;
+}
+
+@media(width<800px) {
+	.leftDash {
+		width: 0%;
+	}
+
+	.rightDash {
+		width: 100%;
+		left: 0;
+	}
+
+	.dashTopbar {
+		font-size: 1rem;
+		padding: 1.5rem .5rem;
+	}
+
+	hr {
+		position: absolute;
+		width: 90%;
+		margin: 0 5%;
+		border: 1px solid #E0EDFF;
+	}
+
+	.content {
+		padding: 1rem;
+	}
+
+	.profileDash {
+		position: relative;
+		margin-left: 0rem;
+		display: flex;
+		flex-wrap: wrap;
+		gap: 2rem;
+		width: 100%;
+	}
+
+	.desc {
+		width: 100%;
+		padding: 1.5rem 1rem;
+		margin-left: 0rem;
+	}
+
+	.description {
+		gap: .5rem;
+	}
+
+	.description div h1 {
+		font-size: 1.5rem;
+	}
+
+	.description div div span {
+		font-size: .7rem;
+	}
+
+	.hackprofile {
+		height: 3.5rem;
+		width: 3.5rem;
+	}
+
+	.love {
+		width: 2rem;
+		height: 2rem;
+	}
+}
