+++ src/modules/Dashboard/modules/LearningCircle/pages/FindCircle.tsx
@@ -30,45 +30,47 @@ const FindCircle = () => {
     }
 
     return (
-    <>
-        <div className={styles.content}>
-            <div className={styles.desc}>
-                <h3>Find your learning circle</h3>
-                <p>Browse and join learning circle around you</p>
-                <div className={styles.ig_search}>
-                    <select name="ig" ref={searchBarRef}>
-                        {igs.map((ig)=><>
-                            <option value={ig.value}>{ig.name}</option>
-                        </>)}
-                    </select>
-                <button onClick={handleSearch}>
-                </button>
-            </div>
-                
-            </div>
-            <img src={imageTop} alt="image"/>
-        </div>
-
-        {(learningCircles)?
-            <div className={styles.container}>
-            {learningCircles.map((circle)=><>
-                <div className={styles.one}>
-                    <h2>{circle.name}</h2>
-                    <p>Team Lead: {circle.lead}</p>
-                    <p>{circle.member_count} Members</p>
-                <div className={styles.join}>
-                    <button>Join</button>
+        <>
+            <div className={styles.content}>
+                <div className={styles.desc}>
+                    <h3>Find your learning circle</h3>
+                    <p>Browse and join learning circle around you</p>
+                    <div className={styles.ig_search}>
+                        <select name="ig" ref={searchBarRef}>
+                            {igs.map(ig => (
+                                <>
+                                    <option value={ig.value}>{ig.name}</option>
+                                </>
+                            ))}
+                        </select>
+                        <button onClick={handleSearch}></button>
+                    </div>
                 </div>
+                <img src={imageTop} alt="image" />
             </div>
-        </>)}
-        </div>
-        :
-        <div className={styles.error_container}>
-            <h1>Found no learning circles :(</h1>
-        </div>
-        }
-    </>
-  )
+
+            {learningCircles ? (
+                <div className={styles.container}>
+                    {learningCircles.map(circle => (
+                        <>
+                            <div className={styles.one}>
+                                <h2>{circle.name}</h2>
+                                <p>Team Lead: {circle.lead}</p>
+                                <p>{circle.member_count} Members</p>
+                                <div className={styles.join}>
+                                    <button>Join</button>
+                                </div>
+                            </div>
+                        </>
+                    ))}
+                </div>
+            ) : (
+                <div className={styles.error_container}>
+                    <h1>Found no learning circles </h1>
+                </div>
+            )}
+        </>
+    );
 }
 
 export default FindCircle
